+++ packages/insomnia/src/ui/insomnia-components/index.ts
@@ -3,7 +3,6 @@ export * from './dropdown';
 export * from './list-group';
 
 export { CardContainer } from './card-container';
-export { Card, type  CardProps } from './card';
 export { MultiSwitch } from './multi-switch';
 export { RadioButtonGroup, type RadioButtonGroupProps } from './radio-button-group';
 export { SvgIcon, type SvgIconProps, type IconEnum, type ThemeEnum, type IconId } from './svg-icon';
